/status?all implemented

+++ SoulissClient/src/it/angelic/soulissclient/model/SoulissTypical.java
@@ -63,8 +63,8 @@ public class SoulissTypical implements Serializable, ISoulissObject, ISoulissTyp
 
 	}
 	
-	public short getOutput() {
-		return typicalDTO.getOutput();
+	public Float getOutput() {
+		return Float.valueOf(typicalDTO.getOutput());
 	}
 
 	@Override
